Close skin egress and nonpersistent hologram storage

+++ plugins/fancyholograms-v2/src/main/java/de/oliver/fancyholograms/HologramManagerImpl.java
@@ -122,7 +122,9 @@ public final class HologramManagerImpl implements HologramManager {
                         }
                     }
 
-                    FancyHolograms.get().getHologramThread().submit(() -> plugin.getHologramStorage().delete(hologram));
+                    if (hologram.getData().isPersistent()) {
+                        FancyHolograms.get().getHologramThread().submit(() -> plugin.getHologramStorage().delete(hologram));
+                    }
                 }
         );
